+++ R/composeVar.R
@@ -46,7 +46,9 @@ composeVar <- function(GADSdat, sourceVars, primarySourceVar, newVar) {
   if(!primarySourceVar %in% sourceVars) stop("'primarySourceVar' must be a name in 'sourceVars'.")
 
   otherSourceVar <- sourceVars[sourceVars != primarySourceVar]
-  suppressWarnings(dat_vec <- extractData(GADSdat)[[primarySourceVar]])
+  suppressWarnings(extracted_dat <- extractData(GADSdat))
+  dat_vec1 <- extracted_dat[[primarySourceVar]]
+  dat_vec2 <- extracted_dat[[otherSourceVar]]
 
   # compare meta (otherwise error?)
   primary_meta <- extractMeta(GADSdat, primarySourceVar)[, c("value", "valLabel", "missings")]
@@ -54,8 +56,10 @@ composeVar <- function(GADSdat, sourceVars, primarySourceVar, newVar) {
   row.names(primary_meta) <- row.names(other_meta) <- NULL
   if(!identical(TRUE, all.equal(primary_meta, other_meta))) stop("Meta data on value level ('value', 'valLabel', 'missings') of the two 'sourceVars' must be identical.")
 
-  # compose
-  comp_var <- ifelse(is.na(dat_vec), yes = GADSdat$dat[[otherSourceVar]],
+  # compose (if primary not missing or both missing => primary)
+  comp_var <- ifelse(is.na(dat_vec1), yes = ifelse(is.na(dat_vec2),
+                                                  yes = GADSdat$dat[[primarySourceVar]],
+                                                  no = GADSdat$dat[[otherSourceVar]]),
                                        no = GADSdat$dat[[primarySourceVar]])
 
   ## put into GADS, add meta